Change : table schema

+++ traindb-core/src/test/resources/datasets/instacart_small/load_mysql.sql
@@ -12,7 +12,7 @@ LOAD DATA LOCAL INFILE 'aisles.csv' INTO TABLE aisles FIELDS TERMINATED BY ',' E
 CREATE TABLE departments (department_id int primary key, department VARCHAR(255));
 LOAD DATA LOCAL INFILE 'departments.csv' INTO TABLE departments FIELDS TERMINATED BY ',' ENCLOSED BY '"' ESCAPED BY '' LINES TERMINATED BY '\n' IGNORE 1 ROWS;
 
-CREATE TABLE orders (order_id int primary key, user_id int, order_number int, order_dow int, order_hour int, day_since_last_order double);
+CREATE TABLE orders (order_id int primary key, user_id int, order_number int, order_dow int, order_hour int, day_since_last_order int);
 LOAD DATA LOCAL INFILE 'orders_small.csv' INTO TABLE orders FIELDS TERMINATED BY ',' ENCLOSED BY '"' ESCAPED BY '' LINES TERMINATED BY '\n' IGNORE 1 ROWS;
 
 CREATE TABLE products (product_id int primary key, product_name VARCHAR(255), aisle_id int NOT NULL, foreign key (aisle_id) references aisles(aisle_id), department_id int NOT NULL, foreign key (department_id) references departments(department_id));
